tif_color: Use 1.0F instead of 1 in Code2V ternary and remove redundant cast
Use a float literal for the fallback value so both branches of the
ternary have the same type, and drop the now-unnecessary outer (float)
cast.

+++ libtiff/tif_color.c
@@ -190,7 +190,7 @@ int TIFFCIELabToRGBInit(TIFFCIELabToRGB *cielab, const TIFFDisplay *display,
 #define ONE_HALF ((int32_t)(1 << (SHIFT - 1)))
 #define Code2V(c, RB, RW, CR)                                                  \
     (((float)((c) - (int32_t)(RB)) * (float)(CR)) /                            \
-     (float)((fabsf((RW) - (RB)) > 0.0F) ? ((RW) - (RB)) : 1))
+     ((fabsf((RW) - (RB)) > 0.0F) ? ((RW) - (RB)) : 1.0F))
 /* !((f)>=(min)) written that way to deal with NaN */
 #define CLAMP(f, min, max)                                                     \
     ((!((f) >= (min))) ? (min) : (f) > (max) ? (max) : (f))
